Updated The Man of the Year Million.
I did some more research and managed to find a stock image of the Pall Mall Budget publication of the text. I also found an older publication of it in the Pall Mall Gazette as well as the caricature in Punch.

This should be the last update as I think I have exhausted the research I can do exclusively online.

It's still the 20th of March as I am writing this so I'm not updating the publication date. Consider this a last minute addition.