Updated The Man of the Year Million
Updated The Man of the Year Million transciption. Found a better source and retranscribed the text. Removed most errors. Found illustrations from Pall Mall Budget version of text and included it in this online transcription. Also provided resources related to the original text. Expanded historical context section and made a greater link to the War of the Worlds.